pyproject.toml
[project]
name = "site-to-site-vpn"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-lib>=2.211.0",
    "constructs>=10.4.2",
    "python-dotenv>=1.1.1",
]

#[project.scripts]
#vpnup = "site_to_site_vpn.cli:main"
[build-system]
requires = ["uv_build>=0.8.5,<0.9.0"]
build-backend = "uv_build"
